iaso_extract_orgunits/requirements.txt
psycopg2-binary
geoalchemy2
topojson
openhexa.toolbox~=2.9.0